requirements.txt
# === LexAssist Backend Dependencies ===

# Web framework
flask>=2.3,<4.0
flask-cors>=4.0,<5.0
gunicorn>=21.2,<23.0

# Supabase
supabase>=2.0,<3.0

# Auth / JWT
PyJWT>=2.8,<3.0

# HTTP
requests>=2.31,<3.0

# Environment
python-dotenv>=1.0,<2.0

# NLP (optional — set ENABLE_INLEGALBERT=true to use)
# transformers>=4.35
# torch>=2.1

# Build tools (ensures pkg_resources is available)
setuptools>=69.0

# Utilities
markupsafe>=2.1
